PR Comments: Move staging object store definitions to their own package.
