+++ flow/connectors/clickhouse/clickhouse.go
@@ -23,6 +23,7 @@ import (
 	"github.com/PeerDB-io/peerdb/flow/generated/protos"
 	"github.com/PeerDB-io/peerdb/flow/internal"
 	peerdb_clickhouse "github.com/PeerDB-io/peerdb/flow/pkg/clickhouse"
+	"github.com/PeerDB-io/peerdb/flow/pkg/objectstore"
 	"github.com/PeerDB-io/peerdb/flow/shared"
 	"github.com/PeerDB-io/peerdb/flow/shared/types"
 )
@@ -79,10 +80,10 @@ func (c *ClickHouseConnector) ValidateCheck(ctx context.Context) error {
 
 	// NOTE: We are skipping staging validation, there is no need to
 	// confirm infrastructure is working at this stage again.
-	// To enable it back, just replace `peerdb_clickhouse.NoStagingValidator`
+	// To enable it back, just replace `objectstore.NoStagingValidator`
 	// with `c.staging.Validate`.
 	if err := peerdb_clickhouse.ValidateClickHousePeer(
-		ctx, c.logger, allowedDomains, c.Config.Host, c.database, peerdb_clickhouse.NoStagingValidator,
+		ctx, c.logger, allowedDomains, c.Config.Host, c.database, objectstore.NoStagingValidator,
 	); err != nil {
 		return err
 	}
